==> notebooks/step3_training.ipynb <==
import time
import pandas as pd
import os
os.chdir('..')
import joblib

# feature columns
X_cols = [
    "hour","dayofweek","is_weekend","month",
    # "lag_24","rolling_24",
    # "electricity_current",
    "airTemperature", "dewTemperature", "windSpeed",
    # "temp_lag_1h","dewTemperature_lag_1h", "windSpeed_lag_1h",
    # "sqft", 
    "sqm", "primaryspaceusage", "site_id", "building_id",
    "Chilledwater", "Hotwater"
]

model_dir = "models_1578_csv"
os.makedirs(model_dir, exist_ok=True)

def load_data(path):
    data_df = pd.read_csv(path)
    return data_df

# TRAINING



data_encoded = "data_1578_csv/train_encode.csv"
data_frame = load_data(data_encoded)

# data_frame[["primaryspaceusage", "site_id", "buiding_id"]].head()
data_frame

import time
import xgboost as xgb

forecast_horizon = 24
times = time.time()
for h in range(forecast_horizon):
    model = xgb.XGBRegressor(
        tree_method="hist",
        n_estimators=500,
        max_depth=10,
        learning_rate=0.01,
        subsample=0.8,
        colsample_bytree=0.8,
        objective="reg:squarederror",
        random_state=42
    )
    
    target_col = f"target_t+{h+1}"
    df_train = data_frame[X_cols + [target_col]].dropna()
    print(f"Training model for horizon {h+1}, training samples: {len(df_train)}")
    X_train = df_train[X_cols]
    y_train = df_train[target_col]
    
    model.fit(X_train, y_train)
    joblib.dump(model, f"{model_dir}/model_hour_{h+1}.pkl")
print("Training time:", time.time() - times)
